+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/Dtos.java
@@ -1907,6 +1907,10 @@ public class Dtos {
       dto.setContainers(asContainerDtos(from.getSequencerPartitionContainers(), includeContainerPartitions, includePoolContents));
     }
 
+    setBoolean(dto::setDataApproved, from.isDataApproved(), true);
+    setId(dto::setDataApproverId, from.getDataApprover());
+    setString(dto::setDataApproverName, maybeGetProperty(from.getDataApprover(), User::getFullName));
+
     return dto;
   }
 
@@ -1966,6 +1970,8 @@ public class Dtos {
     setObject(to::setSequencingParameters, SequencingParameters::new, dto.getSequencingParametersId());
     setObject(to::setSequencingKit, KitDescriptor::new, dto.getSequencingKitId());
     setString(to::setFilePath, dto.getRunPath());
+    setBoolean(to::setDataApproved, dto.isDataApproved(), true);
+    setObject(to::setDataApprover, UserImpl::new, dto.getDataApproverId());
     return to;
   }