+++ sqlstore/src/main/resources/db/migration/V8515__run_approval.sql
@@ -1,0 +1,3 @@
+ALTER TABLE Run ADD COLUMN dataApproved BOOLEAN;
+ALTER TABLE Run ADD COLUMN dataApproverId bigint(20);
+ALTER TABLE Run ADD CONSTRAINT fk_run_approver FOREIGN KEY (dataApproverId) REFERENCES User (userId);